Merged some important changes in the private repo of alm.
- A new tag MAGMOM was implemented for treating spin polarized
systems. The tag is expected to work nicely as the VASP code.
The functions for collinear and noncollinear cases are implemented,
but the noncollinear version is not tested well.
- Fixed a bug in fitting.cpp regarding out-of-range of an array.
This bug appeared in the previous versions when the number of
parameter is larger than the number of training data.
- Improved the way to treat multi-body anharmonic interactions.
- Fixed a bug in calculating the maximum distance for multi-body
interaction clusters.
- Fixed a bug in generating constraints for the translational invariance.
The bug appeared in the previous versions for some special structures
and cutoff radii.
- Fixed a bug in is_incutoff.