Implement FC_ZERO_THR for alm.
FC_ZERO_THR defines the threshold of force constants to be saved in an XML file.
If the absolute value of force constant is smaller than FC_ZERO_THR,
it will NOT be printed out. The default value is FC_ZERO_THR=1.0e-12, which is
the same as the value used in the previous versions.
This parameter was introduced to solve #71. If the harmonic force constants
are calculated using a model potential (e.g., classical FF) where
the interaction becomes zero beyond a certain cutoff raius, the default value
of FC_ZERO_THR may raise a warning when creating a renormalize harmonic FCSXML
using tools/dfc2. This issue may be resolved by using a smaller FC_ZERO_THR,
say FC_ZERO_THR = 1.0e-15.

Note: The force constants that become exactly zero due to symmetry and ASR
constraints will not be printed even when setting FC_ZERO_THR = 0.