Change the way to generate supercell space group operations.
In the current version, the space group operations are first
generated using the primitive cell determined from the input cell
and PRIMCELL tag, and then replicated for the supercell.

When replicating the information, the order of the atoms
given by the users or read from a file will be kept as much as possible.