Fix bugs in printing out HarmonicUnique and CubicUnique.
The sign should be multiplied before printing the fcs values
to the file. As of this change, the output of the HarmonicUnique
and CubicUnique sections will be modified and therefore becomes
incompatible with that generated in the previous versions.

Hence, when using FC2XML and FC3XML tags, combining FC2XML file
generated in the previous version and ALM code as of this version
will result in a strange result (e.g., large fitting error).

To avoid this combination, the code stops when it tries to load
FC2XML and FC3XML files generated in an older version (<=1.4.2).