Implement a function to read force constants from h5
This is used for loading force constants from FC2FIX
and FC3FIX files. In these files, the force constants are
written in the Cartesian basis. Therefore, if FCSYM_BASIS = L,
these force constants need to be converted into the Lattice basis
first, and then be used to generate a constraint matrix (const_fix).

Currently, f the force constant pair of (i,j,k,...) (flattened array)
does not exist in the FC2FIX (FC3FIX) file, the corresponding value
will be set to 0. This is necessary because some force constants
must become exactly 0 due to symmetry and ASR constraints, but they are
not saved in the PREFIX.h5 force constant file to reduce the file
size. To fix this issue, we simply set the unfound pairs to be 0.